rename transformer/pipline to transformer/pipeline
