+++ utensor_cgen/transformer/__init__.py
@@ -3,4 +3,4 @@ from .ns_transformer import *
 from .optimizer import *
 from .quantize import *
 from .cmsis_nn import *
-from .pipline import TransformerPipeline
+from .pipeline import TransformerPipeline
